+++ drel/core/es.py
@@ -5,7 +5,6 @@ from elasticsearch import NotFoundError
 from marshmallow.schema import BaseSchema
 
 from drel.core import config
-from drel.core.config import ELASTIC_SEARCH_RUN_TESTS
 from drel.core.models import FullRequestLog
 from drel.core.schemas import FullRequestLogSchema
 
@@ -17,7 +16,7 @@ def log_to_es(log: FullRequestLog) -> Optional[str]:
 
 
 def write_to_es(doc: Dict) -> Optional[str]:
-    if not ELASTIC_SEARCH_RUN_TESTS:
+    if not config.ELASTIC_SEARCH_RUN_TESTS:
         return None
 
     index = config.INDEX_NAME_GETTER()